djangoapp/backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
distro==1.9.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.99.3
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
gunicorn
whitenoise
          # Postgres (Render/Railway), local SQLite vẫn chạy bình thường
dj-database-url           # tiện đọc DATABASE_URL từ env
python-dotenv             # nếu muốn nạp .env local
psycopg2-binary==2.9.9
